third_party/WebKit/LayoutTests/svg/custom/polyline-setattribute-points-null-expected.txt

CONSOLE ERROR: line 9: Error: <polyline> attribute points: Expected number, "undefined".
